refactor(app): extract headless-run-resume from headless.ts
Moves the execute family (run/resume/watch, retry/recreate/rebase/fork, fix/resolve-conflict) out of headless.ts into headless-run-resume.ts, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
